feat(dsu.md): better impl
