+++ docs/ds/code/dsu/dsu_1.aux1.cpp
@@ -1,0 +1,19 @@
+// https://judge.yosupo.jp/submission/244353
+#include "dsu_1.cpp"
+
+#include <iostream>
+
+int main() {
+  std::cin.tie(nullptr)->std::ios::sync_with_stdio(false);
+  u32 n, q;
+  std::cin >> n >> q;
+  dsu f(n + 1);
+  for (u32 i = 1, t, u, v; i <= q; ++i) {
+    std::cin >> t >> u >> v;
+    if (t == 0)
+      f.merge(u, v);
+    else
+      std::cout << f.same_root(u, v) << '\n';
+  }
+  return 0;
+}
