+++ docs/ds/code/dsu/dsu_1.cpp
@@ -1,64 +1,26 @@
-#include <cassert>
-#include <iostream>
-#include <numeric>
+#include <cstdint>
 #include <vector>
+using u32 = uint32_t;
 
-using namespace std;
+class dsu {
+  // p[i] < 0 时表示 i 为根节点，其对应的子树大小为 -p[i]
+  // p[i] >= 0 时表示 i 不为根节点，其父亲节点的编号为 p[i]
+  std::vector<int> p;
 
-struct dsu {
-  vector<size_t> pa, size, sum;
+ public:
+  // 节点编号从 0 到 sz-1
+  explicit dsu(u32 sz) : p(sz, -1) {}
 
-  explicit dsu(size_t size_)
-      : pa(size_ * 2), size(size_ * 2, 1), sum(size_ * 2) {
-    // size 与 sum 的前半段其实没有使用，只是为了让下标计算更简单
-    iota(pa.begin(), pa.begin() + size_, size_);
-    iota(pa.begin() + size_, pa.end(), size_);
-    iota(sum.begin() + size_, sum.end(), 0);
+  int find(u32 x) { return p[x] < 0 ? (int)x : p[x] = find((u32)p[x]); }
+
+  u32 size(u32 x) { return (u32)-p[(u32)find(x)]; }
+
+  bool same_root(u32 x, u32 y) { return find(x) == find(y); }
+
+  bool merge(u32 x, u32 y) {
+    if ((x = (u32)find(x)) == (y = (u32)find(y))) return false;
+    if (p[x] > p[y]) std::swap(x, y);  // 启发式合并
+    p[x] += p[y], p[y] = (int)x;
+    return true;
   }
-
-  void unite(size_t x, size_t y) {
-    x = find(x), y = find(y);
-    if (x == y) return;
-    if (size[x] < size[y]) swap(x, y);
-    pa[y] = x;
-    size[x] += size[y];
-    sum[x] += sum[y];
-  }
-
-  void move(size_t x, size_t y) {
-    auto fx = find(x), fy = find(y);
-    if (fx == fy) return;
-    pa[x] = fy;
-    --size[fx], ++size[fy];
-    sum[fx] -= x, sum[fy] += x;
-  }
-
-  size_t find(size_t x) { return pa[x] == x ? x : pa[x] = find(pa[x]); }
 };
-
-int main() {
-  size_t n, m, op, x, y;
-  while (cin >> n >> m) {
-    dsu dsu(n + 1);  // 元素范围是 1..n
-    while (m--) {
-      cin >> op;
-      switch (op) {
-        case 1:
-          cin >> x >> y;
-          dsu.unite(x, y);
-          break;
-        case 2:
-          cin >> x >> y;
-          dsu.move(x, y);
-          break;
-        case 3:
-          cin >> x;
-          x = dsu.find(x);
-          cout << dsu.size[x] << ' ' << dsu.sum[x] << '\n';
-          break;
-        default:
-          assert(false);  // not reachable
-      }
-    }
-  }
-}
