+++ docs/ds/examples/dsu/dsu_1.ans
@@ -1,3 +1,4 @@
-3 12
-3 7
-2 8
+0
+1
+0
+1
