+++ docs/ds/examples/dsu/dsu_1.in
@@ -1,8 +1,8 @@
-5 7
+4 7
+1 0 1
+0 0 1
+0 2 3
+1 0 1
 1 1 2
-2 3 4
-1 3 5
-3 4
-2 4 1
-3 4
-3 3
+0 0 2
+1 1 3
