+++ docs/ds/examples/dsu/dsu_2.ans
@@ -1,0 +1,3 @@
+3 12
+3 7
+2 8
