+++ docs/ds/examples/dsu/dsu_2.in
@@ -1,0 +1,8 @@
+5 7
+1 1 2
+2 3 4
+1 3 5
+3 4
+2 4 1
+3 4
+3 3
